+++ sei-ibc-go/testing/mock/ibc_module.go
@@ -6,7 +6,6 @@ import (
 	"strconv"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	capabilitykeeper "github.com/cosmos/cosmos-sdk/x/capability/keeper"
 	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
 
 	channeltypes "github.com/cosmos/ibc-go/v3/modules/core/04-channel/types"
@@ -16,15 +15,16 @@ import (
 
 // IBCModule implements the ICS26 callbacks for testing/mock.
 type IBCModule struct {
-	IBCApp       *MockIBCApp // base application of an IBC middleware stack
-	scopedKeeper capabilitykeeper.ScopedKeeper
+	appModule *AppModule
+	IBCApp    *MockIBCApp // base application of an IBC middleware stack
 }
 
 // NewIBCModule creates a new IBCModule given the underlying mock IBC application and scopedKeeper.
-func NewIBCModule(app *MockIBCApp, scopedKeeper capabilitykeeper.ScopedKeeper) IBCModule {
+func NewIBCModule(appModule *AppModule, app *MockIBCApp) IBCModule {
+	appModule.ibcApps = append(appModule.ibcApps, app)
 	return IBCModule{
-		IBCApp:       app,
-		scopedKeeper: scopedKeeper,
+		appModule: appModule,
+		IBCApp:    app,
 	}
 }
 
@@ -39,7 +39,7 @@ func (im IBCModule) OnChanOpenInit(
 	}
 
 	// Claim channel capability passed back by IBC module
-	if err := im.scopedKeeper.ClaimCapability(ctx, chanCap, host.ChannelCapabilityPath(portID, channelID)); err != nil {
+	if err := im.IBCApp.ScopedKeeper.ClaimCapability(ctx, chanCap, host.ChannelCapabilityPath(portID, channelID)); err != nil {
 		return err
 	}
 
@@ -56,7 +56,7 @@ func (im IBCModule) OnChanOpenTry(
 	}
 
 	// Claim channel capability passed back by IBC module
-	if err := im.scopedKeeper.ClaimCapability(ctx, chanCap, host.ChannelCapabilityPath(portID, channelID)); err != nil {
+	if err := im.IBCApp.ScopedKeeper.ClaimCapability(ctx, chanCap, host.ChannelCapabilityPath(portID, channelID)); err != nil {
 		return "", err
 	}
 
@@ -107,7 +107,7 @@ func (im IBCModule) OnRecvPacket(ctx sdk.Context, packet channeltypes.Packet, re
 
 	// set state by claiming capability to check if revert happens return
 	capName := GetMockRecvCanaryCapabilityName(packet)
-	if _, err := im.scopedKeeper.NewCapability(ctx, capName); err != nil {
+	if _, err := im.IBCApp.ScopedKeeper.NewCapability(ctx, capName); err != nil {
 		// application callback called twice on same packet sequence
 		// must never occur
 		panic(err)
@@ -129,7 +129,7 @@ func (im IBCModule) OnAcknowledgementPacket(ctx sdk.Context, packet channeltypes
 	}
 
 	capName := GetMockAckCanaryCapabilityName(packet)
-	if _, err := im.scopedKeeper.NewCapability(ctx, capName); err != nil {
+	if _, err := im.IBCApp.ScopedKeeper.NewCapability(ctx, capName); err != nil {
 		// application callback called twice on same packet sequence
 		// must never occur
 		panic(err)
@@ -145,7 +145,7 @@ func (im IBCModule) OnTimeoutPacket(ctx sdk.Context, packet channeltypes.Packet,
 	}
 
 	capName := GetMockTimeoutCanaryCapabilityName(packet)
-	if _, err := im.scopedKeeper.NewCapability(ctx, capName); err != nil {
+	if _, err := im.IBCApp.ScopedKeeper.NewCapability(ctx, capName); err != nil {
 		// application callback called twice on same packet sequence
 		// must never occur
 		panic(err)